discountservices/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions
psycopg==3.3.2
psycopg-binary==3.3.2
python-dotenv==1.2.1
requests==2.32.5
selenium==4.39.0
beautifulsoup4==4.14.3
lxml==6.0.2
pandas==2.3.3